#215 fixed mileage data type

+++ backend/app/dtos/filling_dto.py
@@ -9,7 +9,7 @@ from app.dtos.gas_station_dtos import FuelType
 class FillingDTO(BaseModel):
 	license_plate_number: str
 	car_type: str
-	mileage: int
+	mileage: float
 	timestamp: datetime
 	price_per_litre: float
 	litres: float
